Move middleware to its own file
